+++ NEWS.md
@@ -1,5 +1,10 @@
 # spmodel 0.12.0
 
+## Minor Updates
+
+* Added a warning message when `splm()` or `spglm()` converts a non-`POINT` geometry to a `POINT` geometry using `sf::st_centroid()`.
+* Updated a warning message when `splm()` or `spglm()` compute distances using a geographic coordinate system (and not a projected coordinate system).
+
 ## Bug Fixes
 
 * Fixed a bug in the cloud semivariogram (i.e., `esv(..., cloud = TRUE)`) that incorrectly doubled the semivariance.